[Fix] 펑션 제거 및 파일 분리
- struct_CameraPreviewView 파일로 분리
- 프로젝트 파일 사이닝 none으로 변경
- 필요없어진 플래시, 전후면 변경 펑션 제거
