+++ IAteIt/IAteIt/View/Upload/CameraPreviewView.swift
@@ -1,0 +1,39 @@
+//
+//  CameraPreviewView.swift
+//  IAteIt
+//
+//  Created by 유재훈 on 2023/03/06.
+//
+
+import SwiftUI
+import AVFoundation
+
+struct CameraPreviewView: UIViewRepresentable {
+    class VideoPreviewView: UIView {
+        override class var layerClass: AnyClass {
+            AVCaptureVideoPreviewLayer.self
+        }
+        
+        var videoPreviewLayer: AVCaptureVideoPreviewLayer {
+            return layer as! AVCaptureVideoPreviewLayer
+        }
+    }
+    
+    let session: AVCaptureSession
+    
+    func makeUIView(context: Context) -> VideoPreviewView {
+        let view = VideoPreviewView()
+        
+        view.videoPreviewLayer.session = session
+        view.backgroundColor = .black
+        view.videoPreviewLayer.videoGravity = .resizeAspectFill
+        view.videoPreviewLayer.cornerRadius = 0
+        view.videoPreviewLayer.connection?.videoOrientation = .portrait
+        
+        return view
+    }
+    
+    func updateUIView(_ uiView: VideoPreviewView, context: Context) {
+        
+    }
+}
