+++ IAteIt/IAteIt/View/Upload/CameraView.swift
@@ -38,38 +38,6 @@ struct CameraView: View {
     }
 }
 
-
-struct CameraPreviewView: UIViewRepresentable {
-    class VideoPreviewView: UIView {
-        override class var layerClass: AnyClass {
-            AVCaptureVideoPreviewLayer.self
-        }
-        
-        var videoPreviewLayer: AVCaptureVideoPreviewLayer {
-            return layer as! AVCaptureVideoPreviewLayer
-        }
-    }
-    
-    let session: AVCaptureSession
-    
-    func makeUIView(context: Context) -> VideoPreviewView {
-        let view = VideoPreviewView()
-        
-        view.videoPreviewLayer.session = session
-        view.backgroundColor = .black
-        view.videoPreviewLayer.videoGravity = .resizeAspectFill
-        view.videoPreviewLayer.cornerRadius = 0
-        view.videoPreviewLayer.connection?.videoOrientation = .portrait
-        
-        return view
-    }
-    
-    func updateUIView(_ uiView: VideoPreviewView, context: Context) {
-        
-    }
-}
-
-
 struct CameraView_Previews: PreviewProvider {
     static var previews: some View {
         CameraView()
